moved specialization code into a shared folder, as other py versions will also use them
