+++ bytecodemanipulation/data/shared/__init__.py
@@ -1,0 +1,1 @@
+from . import builtin_spec
